==> analysis/01-quality-control.Rmd <==
---
title: "Quality control"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Setup

Set chunk options:

```{r}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/01-quality-control.Rmd/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

Load Bioconductor packages:

```{r}
library(EnsDb.Mmusculus.v79)
library(scater)
library(scran)
library(scuttle)
```

Load CRAN packages:

```{r}
library(eulerr)
library(patchwork)
library(robustbase)
```

Read experiment data:

```{r}
sce <- readRDS("data/sce.rds")
```

Remove version number from gene identifier:

```{r}
ids <- strsplit(rowData(sce)$gene_id, split = ".", fixed = TRUE)

rowData(sce)$gene_id <- sapply(ids, .subset, 1)

rownames(sce) <- rowData(sce)$gene_id
```

Make feature names unique:

```{r}
rownames(sce) <- uniquifyFeatureNames(rowData(sce)$gene_id, rowData(sce)$gene_name)
```

## Exploration

We will begin by exploring the dataset which contains `r ncol(sce)` cells and `r nrow(sce)` genes.

## Metrics

Identify mitochondrial genes:

```{r}
ids <- keys(EnsDb.Mmusculus.v79, keytype = "GENEID", filter = SeqNameFilter("MT"))

idx <- which(rowData(sce)$gene_id %in% ids)
```

Compute QC metrics for each cell:

```{r}
sce <- addPerCellQC(sce, subsets = list(MT = idx))

dat <- makePerCellDF(sce)
```

Inspect the distributions of QC metrics:

```{r}
wrap_plots(
  plotColData(sce, x = "stage", y = "sum") + scale_y_log10() + ggtitle("Total count"),
  plotColData(sce, x = "stage", y = "detected") + scale_y_log10() + ggtitle("Detected features"),
  plotColData(sce, x = "stage", y = "subsets_MT_percent") + ggtitle("MT percent"),
  plotColData(sce, x = "stage", y = "altexps_spike_percent") + ggtitle("Spike percent")
)
```

### Dimensionality reduction {.tabset}

Calculate dimenionality reduction data:

```{r}
set.seed(0110100101)
sce <- computeLibraryFactors(sce)
sce <- logNormCounts(sce)
sce <- runPCA(sce, BSPARAM = BiocSingular::ExactParam())
sce <- runTSNE(sce, dimred = "PCA")
sce <- runUMAP(sce, dimred = "PCA")
```

Inspect dimensionality reduction plots:

#### PCA

```{r}
patchwork::wrap_plots(
  plotPCA(sce, colour_by = "sum", shape_by = "stage") + ggtitle("Total count"),
  plotPCA(sce, colour_by = "detected", shape_by = "stage") + ggtitle("Detected features"),
  plotPCA(sce, colour_by = "subsets_MT_percent", shape_by = "stage") + ggtitle("MT percent"),
  plotPCA(sce, colour_by = "altexps_spike_percent", shape_by = "stage") + ggtitle("Spike percent"),
  ncol = 2
) & coord_fixed(1)
```

#### TSNE

```{r}
patchwork::wrap_plots(
  plotTSNE(sce, colour_by = "sum", shape_by = "stage") + ggtitle("Total count"),
  plotTSNE(sce, colour_by = "detected", shape_by = "stage") + ggtitle("Detected features"),
  plotTSNE(sce, colour_by = "subsets_MT_percent", shape_by = "stage") + ggtitle("MT percent"),
  plotTSNE(sce, colour_by = "altexps_spike_percent", shape_by = "stage") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

#### UMAP

```{r}
patchwork::wrap_plots(
  plotUMAP(sce, colour_by = "sum", shape_by = "stage") + ggtitle("Total count"),
  plotUMAP(sce, colour_by = "detected", shape_by = "stage") + ggtitle("Detected features"),
  plotUMAP(sce, colour_by = "subsets_MT_percent", shape_by = "stage") + ggtitle("MT percent"),
  plotUMAP(sce, colour_by = "altexps_spike_percent", shape_by = "stage") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

## Cell Filtering

We will now perform filtering to remove low-quality cells:

### Manual threshold {.tabset}

The first approach to identifying low-quality cells is to apply manual thresholds on the QC metrics:

```{r}
dat <- DataFrame(
  low_lib_size = sce$sum < 100000,
  low_n_features = sce$detected < 10000,
  high_subsets_MT_percent = sce$subsets_MT_percent > 15,
  high_altexps_spike_percent = sce$altexps_spike_percent > 10
)

sce$manual_discard <- dat$low_lib_size | dat$low_n_features | dat$high_subsets_MT_percent | dat$high_altexps_spike_percent
```

Inspect distribution of QC metrics:

```{r}
wrap_plots(
  plotColData(sce, x = "stage", y = "sum", colour_by = "manual_discard") + scale_y_log10() + ggtitle("Total count"),
  plotColData(sce, x = "stage", y = "detected", colour_by = "manual_discard") + scale_y_log10() + ggtitle("Detected features"),
  plotColData(sce, x = "stage", y = "subsets_MT_percent", colour_by = "manual_discard") + ggtitle("MT percent"),
  plotColData(sce, x = "stage", y = "altexps_spike_percent", colour_by = "manual_discard") + ggtitle("Spike percent")
)
```

Inspect dimensionality reduction plots:

#### PCA

```{r}
patchwork::wrap_plots(
  plotPCA(sce, colour_by = "sum", shape_by = "manual_discard") + ggtitle("Total count"),
  plotPCA(sce, colour_by = "detected", shape_by = "manual_discard") + ggtitle("Detected features"),
  plotPCA(sce, colour_by = "subsets_MT_percent", shape_by = "manual_discard") + ggtitle("MT percent"),
  plotPCA(sce, colour_by = "altexps_spike_percent", shape_by = "manual_discard") + ggtitle("Spike percent"),
  ncol = 2
) & coord_fixed(1)
```

#### TSNE

```{r}
patchwork::wrap_plots(
  plotTSNE(sce, colour_by = "sum", shape_by = "manual_discard") + ggtitle("Total count"),
  plotTSNE(sce, colour_by = "detected", shape_by = "manual_discard") + ggtitle("Detected features"),
  plotTSNE(sce, colour_by = "subsets_MT_percent", shape_by = "manual_discard") + ggtitle("MT percent"),
  plotTSNE(sce, colour_by = "altexps_spike_percent", shape_by = "manual_discard") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

#### UMAP

```{r}
patchwork::wrap_plots(
  plotUMAP(sce, colour_by = "sum", shape_by = "manual_discard") + ggtitle("Total count"),
  plotUMAP(sce, colour_by = "detected", shape_by = "manual_discard") + ggtitle("Detected features"),
  plotUMAP(sce, colour_by = "subsets_MT_percent", shape_by = "manual_discard") + ggtitle("MT percent"),
  plotUMAP(sce, colour_by = "altexps_spike_percent", shape_by = "manual_discard") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

### Adaptive threshold {.tabset}

The next approach is to identify cells that are outliers based on the median absolute deviation for the QC metrics:

```{r}
dat <- quickPerCellQC(colData(sce), batch = sce$stage, sub.fields = c("subsets_MT_percent", "altexps_spike_percent"))

sce$adaptive_discard <- dat$discard
```

Inspect distribution of QC metrics:

```{r}
wrap_plots(
  plotColData(sce, x = "stage", y = "sum", colour_by = "adaptive_discard") + scale_y_log10() + ggtitle("Total count"),
  plotColData(sce, x = "stage", y = "detected", colour_by = "adaptive_discard") + scale_y_log10() + ggtitle("Detected features"),
  plotColData(sce, x = "stage", y = "subsets_MT_percent", colour_by = "adaptive_discard") + ggtitle("MT percent"),
  plotColData(sce, x = "stage", y = "altexps_spike_percent", colour_by = "adaptive_discard") + ggtitle("Spike percent")
)
```

Inspect dimensionality reduction plots:

#### PCA

```{r}
patchwork::wrap_plots(
  plotPCA(sce, colour_by = "sum", shape_by = "adaptive_discard") + ggtitle("Total count"),
  plotPCA(sce, colour_by = "detected", shape_by = "adaptive_discard") + ggtitle("Detected features"),
  plotPCA(sce, colour_by = "subsets_MT_percent", shape_by = "adaptive_discard") + ggtitle("MT percent"),
  plotPCA(sce, colour_by = "altexps_spike_percent", shape_by = "adaptive_discard") + ggtitle("Spike percent"),
  ncol = 2
) & coord_fixed(1)
```

#### TSNE

```{r}
patchwork::wrap_plots(
  plotTSNE(sce, colour_by = "sum", shape_by = "adaptive_discard") + ggtitle("Total count"),
  plotTSNE(sce, colour_by = "detected", shape_by = "adaptive_discard") + ggtitle("Detected features"),
  plotTSNE(sce, colour_by = "subsets_MT_percent", shape_by = "adaptive_discard") + ggtitle("MT percent"),
  plotTSNE(sce, colour_by = "altexps_spike_percent", shape_by = "adaptive_discard") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

#### UMAP

```{r}
patchwork::wrap_plots(
  plotUMAP(sce, colour_by = "sum", shape_by = "adaptive_discard") + ggtitle("Total count"),
  plotUMAP(sce, colour_by = "detected", shape_by = "adaptive_discard") + ggtitle("Detected features"),
  plotUMAP(sce, colour_by = "subsets_MT_percent", shape_by = "adaptive_discard") + ggtitle("MT percent"),
  plotUMAP(sce, colour_by = "altexps_spike_percent", shape_by = "adaptive_discard") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

### Outlier threshold {.tabset}

The last approach is to identify outliers in high-dimensional space based on the QC metrics for each cell:

```{r}
mat <- cbind(log10(sce$sum), log10(sce$detected), sce$subsets_MT_percent, sce$altexps_spike_percent)
adj <- adjOutlyingness(mat, only.outlyingness = TRUE)
out <- isOutlier(adj, type = "higher")
sce$outlier_discard <- out
```

Inspect distribution of QC metrics:

```{r}
wrap_plots(
  plotColData(sce, x = "stage", y = "sum", colour_by = "outlier_discard") + scale_y_log10() + ggtitle("Total count"),
  plotColData(sce, x = "stage", y = "detected", colour_by = "outlier_discard") + scale_y_log10() + ggtitle("Detected features"),
  plotColData(sce, x = "stage", y = "subsets_MT_percent", colour_by = "outlier_discard") + ggtitle("MT percent"),
  plotColData(sce, x = "stage", y = "altexps_spike_percent", colour_by = "outlier_discard") + ggtitle("Spike percent")
)
```

Inspect dimensionality reduction plots:

#### PCA

```{r}
patchwork::wrap_plots(
  plotPCA(sce, colour_by = "sum", shape_by = "outlier_discard") + ggtitle("Total count"),
  plotPCA(sce, colour_by = "detected", shape_by = "outlier_discard") + ggtitle("Detected features"),
  plotPCA(sce, colour_by = "subsets_MT_percent", shape_by = "outlier_discard") + ggtitle("MT percent"),
  plotPCA(sce, colour_by = "altexps_spike_percent", shape_by = "outlier_discard") + ggtitle("Spike percent"),
  ncol = 2
) & coord_fixed(1)
```

#### TSNE

```{r}
patchwork::wrap_plots(
  plotTSNE(sce, colour_by = "sum", shape_by = "outlier_discard") + ggtitle("Total count"),
  plotTSNE(sce, colour_by = "detected", shape_by = "outlier_discard") + ggtitle("Detected features"),
  plotTSNE(sce, colour_by = "subsets_MT_percent", shape_by = "outlier_discard") + ggtitle("MT percent"),
  plotTSNE(sce, colour_by = "altexps_spike_percent", shape_by = "outlier_discard") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

#### UMAP

```{r}
patchwork::wrap_plots(
  plotUMAP(sce, colour_by = "sum", shape_by = "outlier_discard") + ggtitle("Total count"),
  plotUMAP(sce, colour_by = "detected", shape_by = "outlier_discard") + ggtitle("Detected features"),
  plotUMAP(sce, colour_by = "subsets_MT_percent", shape_by = "outlier_discard") + ggtitle("MT percent"),
  plotUMAP(sce, colour_by = "altexps_spike_percent", shape_by = "outlier_discard") + ggtitle("Spike percent"),
  ncol = 2
) & theme(aspect.ratio = 1)
```

### Comparison

Compare the number of cells selected between methods using an Euler diagram:

```{r}
mat <- cbind(Keep = TRUE, Manual = sce$manual_discard, Adaptive = sce$adaptive_discard, Outlier = sce$outlier_discard)
fit <- euler(mat, shape = "circle")
col <- RColorBrewer::brewer.pal(n = 4, name = "Pastel2")
plot(fit, fill = col, quantities = TRUE)
```

### Selection

We are going to use the adaptive filtering to remove low quality cells:

```{r}
sce <- sce[, !sce$adaptive_discard]
```

## Gene filtering

Compute QC metrics for each gene:

```{r}
sce <- addPerFeatureQC(sce)

dat <- as.data.frame(rowData(sce))
```

### Expression rate

Plot the distribution of log-means across all genes:

```{r}
ggplot(dat, aes(mean)) + 
  geom_histogram(bins = 100, colour = "black", fill = "grey") + 
  scale_x_log10(n.breaks = 10) +
  theme_bw()
```

Define lowly expressed genes using a mean threshold:

```{r}
dat$low_mean <- dat$mean < 1
```

Tabulate the number of lowly expressed genes:

```{r}
knitr::kable(table(dat$low_mean), col.names = c("Lowly expressed", "Count"))
```

### Detection rate

Plot the percentage of expressing cells against the log-mean expression for each gene:

```{r}
ggplot(dat, aes(mean, detected)) +
  geom_point(alpha = 0.1, colour = "grey") +
  geom_smooth(colour = "black", se = FALSE) +
  scale_x_log10(n.breaks = 10) +
  theme_bw()
```

Define lowly detected genes as having a minimum of 1 read in at least 1 percent of cells:

```{r}
dat$low_detected <- !rowSums(counts(sce) >= 1) >= (ncol(sce) * 0.01)
```

Tabulate the number of lowly detected genes:

```{r}
knitr::kable(table(dat$low_detected), col.names = c("Lowly detected", "Count"))
```

### Comparison

Compare the number of genes selected between methods using an Euler diagram:

```{r}
mat <- cbind(Keep = TRUE, Mean = dat$low_mean, Freq = dat$low_detected)
fit <- euler(mat, shape = "circle")
col <- RColorBrewer::brewer.pal(n = 3, name = "Pastel2")
plot(fit, fill = col, quantities = TRUE)
```

### Selection

Bookmark low abundance and frequency genes (no filtering is performed because of data integration):

```{r}
rowData(sce) <- cbind(rowData(sce), dat[, c("low_mean", "low_detected")])
```

## Summary

After quality control we have a dataset with `r ncol(sce)` cells and `r nrow(sce)` genes.

### Output

Save experiment object to disk:

```{r}
saveRDS(sce, "output/01-quality-control.rds")
```

### Session

Print session information:

```{r}
sessionInfo()
```
